+++ CHANGELOG.md
@@ -8,12 +8,15 @@ This release has some backwards incompatible breaks, including:
 
 - `Phly\Http\Request` no longer accepts an HTTP protocol version as a constructor argument. Use `setProtocolVersion()` instead.
 - `Phly\Http\Request` now uses `php://memory` as the default body stream. (`IncomingRequest` uses `php://input` as the default stream.)
-
+- `Phly\Http\RequestFactory` has been renamed to `Phly\Http\IncomingRequestFactory`
+  - It also now expects an `IncomingRequestInterface` if passed a request object to populate.
+  
 ### Added
 
 - `Phly\Http\MessageTrait::setProtocolVersion($version)`, per changes in PSR-7 (this is now defined in the `MessageInterface`).
 - Note in `Phly\Http\Stream::read()`'s `@return` annotation indicating that it can also return boolean `false`.
 - `Phly\Http\IncomingRequest`, which implements `Psr\Http\Message\IncomingRequestInterface` and provides a server-side request implementation with accessors for each type of request datum typically accessed (cookies, matched path parameters, query string arguments, body parameters, and upload file information). It uses `php://input` as the default body stream.
+- `Phly\Http\IncomingRequestFactory` (which replaces `Phly\Http\RequestFactory`)
 
 ### Deprecated
 
@@ -21,7 +24,7 @@ This release has some backwards incompatible breaks, including:
 
 ### Removed
 
-- Nothing.
+- `Phly\Http\RequestFactory` (it is now `Phly\Http\IncomingRequestFactory`)
 
 ### Fixed